pyproject.toml
[project]
name = "mindful"
version = "0.1.0"
description = "Mindful is a Note Taking application with a focus on simplicity and ease of use."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi",
    "beautifulsoup4",
    "chroma>=0.2.0",
    "chromadb",
    "fastapi>=0.115.8",
    "langchain>=0.3.18",
    "langchain-chroma>=0.2.1",
    "langchain-community>=0.3.17",
    "langchain-core>=0.3.34",
    "langchain-groq>=0.2.4",
    "langchain-nomic>=0.1.4",
    "langchain-ollama>=0.2.3",
    "motor>=3.7.0",
    "passlib>=1.7.4",
    "pydantic>=2.10.6",
    "pypdf>=5.3.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "unstructured>=0.16.20",
    "uvicorn>=0.34.0",
]
